Create English-to-lolspeakVH
For this project, I built an English-to-Lolcat translation engine. It contains a glossary of more than 2,200 words as a basis (English words that don’t have a Lolcat
equivalent are left as they are).

This program reads a plain-text file in English (.txt extension), translates it,
and produce another text file in Lolcat, regardless of the length of the text.